refactor some pass
