+++ frontends/torch-frontend/torch-frontend/include/torch-frontend/Conversion/Passes.h
@@ -20,7 +20,6 @@
 
 #include "torch-frontend/Conversion/ConvertTorchToCustomCall.h"
 #include "torch-frontend/Conversion/ConvertTorchToStablehloExt.h"
-#include "torch-frontend/Conversion/FuseOpOnTorch.h"
 
 namespace mlir {
