+++ frontends/torch-frontend/torch-frontend/include/torch-frontend/Transforms/Passes.h
@@ -21,8 +21,8 @@
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "torch-frontend/Transforms/CanonicalizeExt.h"
 #include "torch-frontend/Transforms/EliminateUselessOp.h"
+#include "torch-frontend/Transforms/FuseOpOnTorch.h"
 #include "torch-frontend/Transforms/RewriteCustomOp.h"
-#include "torch-frontend/Transforms/RewriteToByteIREntryPoint.h"
 #include "torch-frontend/Transforms/UnpackPublicFunctionReturn.h"
 
 namespace mlir {
