+++ frontends/torch-frontend/torch-frontend/lib/Conversion/CMakeLists.txt
@@ -1,15 +1,9 @@
-set(LLVM_TARGET_DEFINITIONS FuseOpOnTorchPattern.td)
-mlir_tablegen(FuseOpOnTorchPattern.inc -gen-rewriters)
-add_public_tablegen_target(FuseOpOnTorchPatternIncGen)
-
 add_mlir_library(TorchFrontendConversion
   ConvertTorchToCustomCall.cpp
   ConvertTorchToStablehloExt.cpp
-  FuseOpOnTorch.cpp
 
   DEPENDS
   TorchFrontendConversionPassIncGen
-  FuseOpOnTorchPatternIncGen
 
   LINK_COMPONENTS
   Core
@@ -28,4 +22,3 @@ add_mlir_library(TorchFrontendConversion
   TorchMLIRTorchConversionPasses
   TorchMLIRTorchUtils
 )
-target_include_directories(TorchFrontendConversion PUBLIC ${CMAKE_CURRENT_BINARY_DIR})
