--- frontends/torch-frontend/torch-frontend/lib/Transforms/RewriteToByteIREntryPoint.cpp
@@ -1,42 +1,0 @@
-//===- RewriteToByteIREntryPoint.cpp --------------------------*--- C++ -*-===//
-//
-// Copyright 2022 ByteDance Ltd. and/or its affiliates. All rights reserved.
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-//===----------------------------------------------------------------------===//
-
-#include "torch-frontend/Transforms/RewriteToByteIREntryPoint.h"
-#include "mlir/IR/BuiltinOps.h"
-
-#include "./PassDetail.h"
-
-using namespace mlir;
-
-namespace {
-
-struct RewriteToByteIREntryPointPass
-    : public RewriteToByteIREntryPointBase<RewriteToByteIREntryPointPass> {
-  RewriteToByteIREntryPointPass() = default;
-  void runOnOperation() override final {
-    ModuleOp m = getOperation();
-    MLIRContext *ctx = &getContext();
-    // TODO
-  }
-};
-
-} // namespace
-
-std::unique_ptr<OperationPass<ModuleOp>>
-mlir::createRewriteToByteIREntryPoint() {
-  return std::make_unique<RewriteToByteIREntryPointPass>();
-}
